refactor: remove old files and organised sections
