--- src/components/Links.module.css
@@ -1,35 +1,0 @@
-.links-container {
-  height: calc(100% - 8.35rem);
-  overflow-y: auto;
-  padding: 0 0.5rem;
-
-  /* Hide scrollbar for IE, Edge and Firefox */
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none; /* Firefox */
-}
-
-.links-container::-webkit-scrollbar {
-  /* Hide scrollbar for Chrome, Safari and Opera */
-  display: none;
-}
-
-.links-button {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background-color: hsl(265, 100%, 91%);
-  border-radius: 5px;
-  padding: 0.5rem;
-}
-
-.links-button:hover {
-  background-color: hsl(265, 100%, 88%);
-}
-
-.title {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding-right: 0.5rem;
-}
